Fixes for CI
